Split fnConfig.cpp into separate files
