+++ include/version.h
@@ -10,8 +10,8 @@
 #define FN_VERSION_MAJOR 0
 #define FN_VERSION_MINOR 5
 
-#define FN_VERSION_BUILD "111eb575"
+#define FN_VERSION_BUILD "987fb160"
 
-#define FN_VERSION_DATE "2022-11-19 11:08:58"
+#define FN_VERSION_DATE "2022-11-21 19:27:32"
 
-#define FN_VERSION_FULL "0.5.111eb575"
+#define FN_VERSION_FULL "0.5.987fb160"
