+++ lib/config/fnConfig.h
@@ -15,6 +15,14 @@
 
 #define HSIO_INVALID_INDEX -1
 
+#define CONFIG_FILENAME "/fnconfig.ini"
+#define CONFIG_FILEBUFFSIZE 2048
+
+#define CONFIG_DEFAULT_SNTPSERVER "pool.ntp.org"
+
+#define PHONEBOOK_CHAR_WIDTH 12 
+
+
 class fnConfig
 {
 public:
